(#191) Rename files with JSX as .jsx
- This also adds a linter rule that should be in standards.

Closes #191
