+++ src/components/atomic/RemovableTag/__tests__/RemovableTag.tests.jsx
