+++ src/utils/getLocationInfoFromSites.jsx
