+++ src/views/Disease/__tests__/Disease.Error.test.jsx
@@ -3,20 +3,19 @@ import React from 'react';
 import { MemoryRouter } from 'react-router';
 
 import Disease from '../Disease';
-import ErrorBoundary from '../../ErrorBoundary/ErrorBoundary';
-import { useStateValue } from '../../../store/store.js';
+import { useStateValue } from '../../../store/store';
 import { MockAnalyticsProvider } from '../../../tracking';
 import { useAppPaths } from '../../../hooks/routing';
 import { useCtsApi } from '../../../hooks/ctsApiSupport/useCtsApi';
 
-jest.mock('../../../store/store.js');
+jest.mock('../../../store/store');
 jest.mock('../../../hooks/routing');
 jest.mock('../../../hooks/ctsApiSupport/useCtsApi');
 
 jest.mock('react-router', () => ({
 	...jest.requireActual('react-router'),
 	useParams: () => ({
-		foo: 'bar',
+		codeOrPurl: 'breast-cancer',
 	}),
 }));
 
@@ -25,7 +24,7 @@ describe('<Disease />', () => {
 		jest.clearAllMocks();
 	});
 
-	it('should throw on unknown param', async () => {
+	it('should render <ErrorPage /> component', async () => {
 		const basePath = '/';
 		const canonicalHost = 'https://www.cancer.gov';
 		const data = [
@@ -108,7 +107,7 @@ describe('<Disease />', () => {
 
 		const routeParamMap = [
 			{
-				paramName: 'chicken',
+				paramName: 'codeOrPurl',
 				textReplacementKey: 'disease',
 				type: 'listing-information',
 			},
@@ -117,20 +116,18 @@ describe('<Disease />', () => {
 		await act(async () => {
 			render(
 				<MockAnalyticsProvider>
-					<ErrorBoundary>
-						<MemoryRouter initialEntries={['/C4872']}>
-							<Disease
-								routeParamMap={routeParamMap}
-								routePath={redirectPath}
-								data={data}
-							/>
-						</MemoryRouter>
-					</ErrorBoundary>
+					<MemoryRouter initialEntries={['/C4872']}>
+						<Disease
+							routeParamMap={routeParamMap}
+							routePath={redirectPath}
+							data={data}
+						/>
+					</MemoryRouter>
 				</MockAnalyticsProvider>
 			);
 		});
 
-		expect(useCtsApi).not.toHaveBeenCalled();
+		expect(useCtsApi).toHaveBeenCalled();
 
 		expect(
 			screen.getByText('An error occurred. Please try again later.')
