+++ src/views/Disease/__tests__/Disease.NoTrialsFoundRedirect.test.jsx
@@ -3,13 +3,13 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { MemoryRouter, useLocation } from 'react-router';
 import Disease from '../Disease';
-import { useStateValue } from '../../../store/store.js';
+import { useStateValue } from '../../../store/store';
 import { MockAnalyticsProvider } from '../../../tracking';
 import { useCtsApi } from '../../../hooks/ctsApiSupport/useCtsApi';
 import { getClinicalTrials } from '../../../services/api/actions/getClinicalTrials';
 
 jest.mock('../../../hooks/ctsApiSupport/useCtsApi');
-jest.mock('../../../store/store.js');
+jest.mock('../../../store/store');
 
 jest.mock('react-router', () => ({
 	...jest.requireActual('react-router'), // use actual for all non-hook parts
