+++ src/views/Disease/__tests__/Disease.TrialType.test.jsx
@@ -1,53 +1,59 @@
-import { act, render, screen } from '@testing-library/react';
+import { act, fireEvent, render, screen } from '@testing-library/react';
 import React from 'react';
 import { MemoryRouter } from 'react-router';
 
 import Disease from '../Disease';
-import { useStateValue } from '../../../store/store.js';
+import { useStateValue } from '../../../store/store';
 import { MockAnalyticsProvider } from '../../../tracking';
+import { getClinicalTrials } from '../../../services/api/actions/getClinicalTrials';
 import { useAppPaths } from '../../../hooks/routing';
 import { useCtsApi } from '../../../hooks/ctsApiSupport/useCtsApi';
 
-jest.mock('../../../store/store.js');
+jest.mock('../../../store/store');
 jest.mock('../../../hooks/routing');
 jest.mock('../../../hooks/ctsApiSupport/useCtsApi');
 
-jest.mock('react-router', () => ({
-	...jest.requireActual('react-router'),
-	useParams: () => ({
-		codeOrPurl: 'breast-cancer',
-	}),
-}));
+const fixturePath = `/v2/trials`;
+const trastuzumabFile = `trastuzumab-response.json`;
 
 describe('<Disease />', () => {
 	afterEach(() => {
 		jest.clearAllMocks();
 	});
 
-	it('should render <ErrorPage /> component', async () => {
+	useCtsApi.mockReturnValue({
+		error: false,
+		loading: false,
+		aborted: false,
+		payload: {
+			total: 0,
+			data: [],
+		},
+	});
+
+	it('should render <ResultsList /> component', async () => {
 		const basePath = '/';
 		const canonicalHost = 'https://www.cancer.gov';
 		const data = [
 			{
-				conceptId: ['C4872'],
-				name: {
-					label: 'Breast Cancer',
-					normalized: 'breast cancer',
-				},
+				conceptId: ['C4872', 'C118809'],
+				name: { label: 'Breast Cancer', normalized: 'breast cancer' },
 				prettyUrlName: 'breast-cancer',
 			},
+			{ prettyUrlName: 'treatment', idString: 'treatment', label: 'Treatment' },
 		];
 		const detailedViewPagePrettyUrlFormatter = '/clinicaltrials/{{nci_id}}';
 		const title = 'NCI Clinical Trials';
 		const trialListingPageType = 'Disease';
 		const dynamicListingPatterns = {
 			Disease: {
-				browserTitle: '{{disease_name}} Clinical Trials',
+				browserTitle: '{{disease_label}} Clinical Trials',
 				introText:
-					'<p>Clinical trials are research studies that involve people. The clinical trials on this list are for {{disease_normalized}}.</p>',
-				metaDescription: 'Find clinical trials for {{disease_normalized}}.',
+					'<p>Clinical trials are research studies that involve people. The clinical trials on this list are for {{disease_normalized}}. All trials on the list are NCI-supported clinical trials, which are sponsored or otherwise financially supported by NCI.</p><p>NCI’s <a href="/about-cancer/treatment/clinical-trials/what-are-trials">basic information about clinical trials</a> explains the types and phases of trials and how they are carried out. Clinical trials look at new ways to prevent, detect, or treat disease. You may want to think about taking part in a clinical trial. Talk to your doctor for help in deciding if one is right for you.</p>',
+				metaDescription:
+					'NCI supports clinical trials studying new and more effective ways to detect and treat cancer. Find clinical trials for {{disease_normalized}}.',
 				noTrialsHtml:
-					'<p>There are currently no available trials for {{disease_normalized}}.</p>',
+					'<p>There are no NCI-supported clinical trials for {{disease_normalized}} at this time. You can try a <a href="/about-cancer/treatment/clinical-trials/search">new search</a> or <a href="/contact">contact our Cancer Information Service</a> to talk about options for clinical trials.</p>',
 				pageTitle: '{{disease_label}} Clinical Trials',
 			},
 			DiseaseTrialType: {
@@ -74,6 +80,7 @@ describe('<Disease />', () => {
 					'{{trial_type_label}} Clinical Trials for {{disease_label}} Using {{intervention_label}}',
 			},
 		};
+		const response = getFixture(`${fixturePath}/${trastuzumabFile}`);
 
 		useStateValue.mockReturnValue([
 			{
@@ -85,22 +92,18 @@ describe('<Disease />', () => {
 				itemsPerPage: 25,
 				title,
 				trialListingPageType,
-				apiClients: {
-					clinicalTrialsSearchClient: true,
-				},
-				language: 'en',
 			},
 		]);
 
 		useAppPaths.mockReturnValue({
-			codeOrPurlPath: '/:codeOrPurl',
+			CodeOrPurlWithTypePath: () => '/:codeOrPurl/:type',
 		});
 
 		useCtsApi.mockReturnValue({
-			error: new Error('Bad Mojo'),
+			error: false,
 			loading: false,
 			aborted: false,
-			payload: null,
+			payload: response,
 		});
 
 		const redirectPath = () => '/notrials';
@@ -111,12 +114,17 @@ describe('<Disease />', () => {
 				textReplacementKey: 'disease',
 				type: 'listing-information',
 			},
+			{
+				paramName: 'type',
+				textReplacementKey: 'trial_type',
+				type: 'trial-type',
+			},
 		];
 
 		await act(async () => {
 			render(
 				<MockAnalyticsProvider>
-					<MemoryRouter initialEntries={['/C4872']}>
+					<MemoryRouter initialEntries={['/C4872/treatment']}>
 						<Disease
 							routeParamMap={routeParamMap}
 							routePath={redirectPath}
@@ -127,10 +135,38 @@ describe('<Disease />', () => {
 			);
 		});
 
+		const requestFilters = {
+			'diseases.nci_thesaurus_concept_id': ['C4872', 'C118809'],
+			primary_purpose: 'treatment',
+		};
+		const requestQuery = getClinicalTrials({
+			from: 0,
+			requestFilters,
+			size: 25,
+		});
+
+		expect(useCtsApi.mock.calls[0][0]).toEqual(requestQuery);
 		expect(useCtsApi).toHaveBeenCalled();
 
 		expect(
-			screen.getByText('An error occurred. Please try again later.')
+			screen.getByText('Treatment Clinical Trials for Breast Cancer')
 		).toBeInTheDocument();
+		expect(
+			screen.getByText(
+				'Clinical trials are research studies that involve people. The clinical trials on this list are for breast cancer treatment. All trials on the list are NCI-supported clinical trials, which are sponsored or otherwise financially supported by NCI.'
+			)
+		).toBeInTheDocument();
+
+		// Navigate to page 2 with next pager item. Confirm currently active page on top and bottom is 2
+		fireEvent.click(screen.getAllByRole('button', { name: 'next page' })[0]);
+
+		expect(screen.getAllByRole('button', { name: 'page 2' })[0]).toHaveClass(
+			'pager__button active',
+			{ exact: true }
+		);
+		expect(screen.getAllByRole('button', { name: 'page 2' })[1]).toHaveClass(
+			'pager__button active',
+			{ exact: true }
+		);
 	});
 });
