+++ src/views/Intervention/__tests__/Intervention.BadParam.test.jsx
@@ -3,31 +3,39 @@ import React from 'react';
 import { MemoryRouter } from 'react-router';
 
 import Intervention from '../Intervention';
-import { useStateValue } from '../../../store/store.js';
+import ErrorBoundary from '../../ErrorBoundary/ErrorBoundary';
+import { useStateValue } from '../../../store/store';
 import { MockAnalyticsProvider } from '../../../tracking';
 import { useAppPaths } from '../../../hooks/routing';
 import { useCtsApi } from '../../../hooks/ctsApiSupport/useCtsApi';
 
-jest.mock('../../../store/store.js');
+jest.mock('../../../store/store');
 jest.mock('../../../hooks/routing');
 jest.mock('../../../hooks/ctsApiSupport/useCtsApi');
+
+jest.mock('react-router', () => ({
+	...jest.requireActual('react-router'),
+	useParams: () => ({
+		foo: 'bar',
+	}),
+}));
 
 describe('<Intervention />', () => {
 	afterEach(() => {
 		jest.clearAllMocks();
 	});
 
-	it('should render <ErrorPage> Component', async () => {
+	it('should throw on unknown param', async () => {
 		const basePath = '/';
 		const canonicalHost = 'https://www.cancer.gov';
 		const data = [
 			{
-				conceptId: ['C1647'],
+				conceptId: ['C4872'],
 				name: {
-					label: 'Trastuzumab',
-					normalized: 'trastuzumab',
+					label: 'Breast Cancer',
+					normalized: 'breast cancer',
 				},
-				prettyUrlName: 'trastuzumab',
+				prettyUrlName: 'breast-cancer',
 			},
 		];
 		const detailedViewPagePrettyUrlFormatter = '/clinicaltrials/{{nci_id}}';
@@ -54,12 +62,12 @@ describe('<Intervention />', () => {
 				detailedViewPagePrettyUrlFormatter,
 				dynamicListingPatterns,
 				itemsPerPage: 25,
-				language: 'en',
 				title,
 				trialListingPageType,
 				apiClients: {
 					clinicalTrialsSearchClient: true,
 				},
+				language: 'en',
 			},
 		]);
 
@@ -68,17 +76,18 @@ describe('<Intervention />', () => {
 		});
 
 		useCtsApi.mockReturnValue({
-			error: false,
+			error: new Error('Bad Mojo'),
 			loading: false,
 			aborted: false,
 			payload: null,
 		});
 
 		const redirectPath = () => '/notrials';
+
 		const routeParamMap = [
 			{
-				paramName: 'codeOrPurl',
-				textReplacementKey: 'disease',
+				paramName: 'chicken',
+				textReplacementKey: 'Intervention',
 				type: 'listing-information',
 			},
 		];
@@ -86,22 +95,21 @@ describe('<Intervention />', () => {
 		await act(async () => {
 			render(
 				<MockAnalyticsProvider>
-					<MemoryRouter initialEntries={['/trastuzumab']}>
-						<Intervention
-							routeParamMap={routeParamMap}
-							routePath={redirectPath}
-							data={data}
-						/>
-					</MemoryRouter>
+					<ErrorBoundary>
+						<MemoryRouter initialEntries={['/C4872']}>
+							<Intervention
+								routeParamMap={routeParamMap}
+								routePath={redirectPath}
+								data={data}
+							/>
+						</MemoryRouter>
+					</ErrorBoundary>
 				</MockAnalyticsProvider>
 			);
 		});
 
-		expect(useCtsApi).toHaveBeenCalled();
+		expect(useCtsApi).not.toHaveBeenCalled();
 
-		expect(
-			screen.getByText('Clinical Trials Using Trastuzumab')
-		).toBeInTheDocument();
 		expect(
 			screen.getByText('An error occurred. Please try again later.')
 		).toBeInTheDocument();
